// modules/contracts/interfaces/contractStatus.ts
export type ContractStatus =
  | 'active'
  | 'processed'
  | 'recoverable_cancellation'
  | 'expired'
  | 'pending_signature'
  | 'incident'
  | 'administrative_incident'
  | 'KO'
  | 'not_commissionable'
  | 'pending_activation'
  | 'pending_verification'
  | 'scoring'
  | 'scoring_KO'
  | 'temporary'
  | 'processed_very_KO'